requirements.txt
Django
djangorestframework
icalendar
requests

